+++ Src/FluentAssertions/Specialized/DelegateAssertions.cs
@@ -31,6 +31,22 @@ public abstract class DelegateAssertions<TDelegate, TAssertions> : DelegateAsser
     }
 
     /// <summary>
+    /// Asserts that the current <see cref="Delegate" /> throws any exception.
+    /// </summary>
+    /// <param name="because">
+    /// (Optional)
+    /// A formatted phrase as is supported by <see cref="string.Format(string,object[])" /> explaining why the assertion is needed. If
+    /// the phrase does not start with the word <i>because</i>, it is prepended automatically.
+    /// </param>
+    /// <param name="becauseArgs">
+    /// Zero or more objects to format using the placeholders in <paramref name="because" />.
+    /// </param>
+    public ExceptionAssertions<Exception> Throw([StringSyntax("CompositeFormat")] string because = "", params object[] becauseArgs)
+    {
+        return Throw<Exception>(because, becauseArgs);
+    }
+
+    /// <summary>
     /// Asserts that the current <see cref="Delegate" /> throws an exception of type <typeparamref name="TException"/>.
     /// </summary>
     /// <param name="because">